requirements.txt
python-telegram-bot==20.8
pydub
translate
gtts
SpeechRecognition
python-dotenv==1.1.1